fix(tests): resolve compilation errors in existing test files
Fix 14 test files to compile against current production APIs:
- Add @MainActor isolation for MainActor-isolated dependencies
- Update renamed types (PeekDirection, TikTokRallyPlayerView, etc.)
- Fix API signatures (TrackedBall init, VideoMetadata init, etc.)
- Fix Swift syntax (tuple destructuring, string interpolation, closures)
- Add TrackedBallTestHelper for shared test convenience init
- Fix tearDown ordering in DebugVideoExporterTests (nil-after-use crash)

Zero production code changes. All 23 test suites now compile and execute.

+++ BumpSetCutTests/Data/Storage/MediaStoreSearchTests.swift
@@ -61,7 +61,7 @@ final class MediaStoreSearchTests: XCTestCase {
         ]
         
         for (fileName, folderPath, customName, fileSize) in testVideos {
-            let testURL = createTestVideoURL(fileName: fileName, fileSize: fileSize)
+            let testURL = createTestVideoURL(fileName: fileName, fileSize: Int64(fileSize))
             _ = mediaStore.addVideo(at: testURL, toFolder: folderPath, customName: customName)
         }
     }
